Remove chunk option on VSGAN
The chunk option shouldn't ever be realistically used. If you have a scenario in which it would be used, and you actually understand why it's bad and still wish to, then you most likely do not need to use a boilerplate script like VSMPEG and are well capable without it.