Remove "video_in" as input, (aka mpv input)
MPV cannot properly be supported with the underlying pvsfunc code as MPV does not expose the original file path. It instead renders the video data and passes said data to VS. Thats rank if you want to do your own indexing and fame serving, which pvsfunc needs for accurate and efficient deinterlacing and vfr to cfr.