Set d2v_vst_vfr_mode to the default of false
This option shouldnt be set to true right now as its broken on a fair few sources (it will cause an exception). You can try it if you wish, regardless an update has been made since.

The update however, is a big change which actually deprecates PSourcer and PDeinterlacer completely for MPEG-1 and MPEG-2 sources. Due to the code not being updated for AVC/other codecs in general, I cannot yet update the code to use the new PD2V (which has the fix).